Moved basic validator to base and added jar-compilation for base
