+++ plugins/de/opendiabetes/vault/plugin/importer/googlefit/GoogleFitCSVValidator.java
@@ -17,7 +17,7 @@
 package de.opendiabetes.vault.plugin.importer.googlefit;
 
 import com.csvreader.CsvReader;
-import de.opendiabetes.vault.plugin.importer.validator.CSVValidator;
+import de.opendiabetes.vault.plugin.fileimporter.validator.CSVValidator;
 import de.opendiabetes.vault.plugin.util.TimestampUtils;
 
 import java.io.File;
