+++ plugins/de/opendiabetes/vault/plugin/importer/mysugr/MySugrCSVValidator.java
@@ -18,7 +18,7 @@ package de.opendiabetes.vault.plugin.importer.mysugr;
 
 import com.csvreader.CsvReader;
 import de.opendiabetes.vault.container.VaultEntryAnnotation;
-import de.opendiabetes.vault.plugin.importer.validator.CSVValidator;
+import de.opendiabetes.vault.plugin.fileimporter.validator.CSVValidator;
 import de.opendiabetes.vault.plugin.util.TimestampUtils;
 
 import java.io.IOException;
