--- plugins/de/opendiabetes/vault/plugin/importer/validator/package-info.java
@@ -1,4 +1,0 @@
-/**
- * This class contains the validators used by importers.
- */
-package de.opendiabetes.vault.plugin.importer.validator;
