+++ src/de/opendiabetes/vault/plugin/fileimporter/validator/package-info.java
@@ -1,0 +1,4 @@
+/**
+ * This class contains the abstract validator used by importers.
+ */
+package de.opendiabetes.vault.plugin.fileimporter.validator;
